# transportManager/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Transport Management's API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/manager/user": {
            "get": {
                "tags": [
                    "Manager"
                ],
                "summary": "all users",
                "operationId": "d1d7505016efa9800d5c8961e329385d",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/manager/user/add": {
            "post": {
                "tags": [
                    "Manager"
                ],
                "operationId": "083eae25636b9980a8496a3342aad72f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "unAutosize",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/manager/login": {
            "get": {
                "tags": [
                    "Manager"
                ],
                "summary": "Logs user into the system",
                "description": "",
                "operationId": "loginUser",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The user name for login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The password for login in clear text",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "headers": {
                            "X-Rate-Limit": {
                                "description": "calls per hour allowed by the user",
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
...
        "/manager/constantType": {
            "post": {
                "tags": [
                    "Manager"
                ],
                "operationId": "cc40734a4322e554ac6492e9420d369d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "title_p"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "title_p": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "unAutosize",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/manager/constantType/{id}": {
            "put": {
                "tags": [
                    "Manager"
                ],
                "operationId": "c712d253d483a508f7bdae489fcf3461",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "title_p"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "title_p": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "unAutosize",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/manager/cities": {
            "get": {
                "tags": [
                    "Manager"
                ],
                "summary": "all city",
                "operationId": "60b422a3aa6dbdd883f374d03464b108",
                "parameters": [
                    {
                        "name": "city_name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/manager/constant/type": {
            "get": {
                "tags": [
                    "Manager"
                ],
                "summary": "all constants type",
                "operationId": "267c562de9b6247ac960e4b3af18485c",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/manager/constants": {
            "get": {
                "tags": [
                    "Manager"
                ],
                "summary": "all constant",
                "operationId": "86639ede943813cf4b02c4fbc7e979da",
                "parameters": [
                    {
                        "name": "type_name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/manager/crmLoaders": {
            "get": {
                "tags": [
                    "Manager"
                ],
                "summary": "all constant",
                "operationId": "3024559fadb1705b04acf015f7ab67b2",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}